Interpolation working
This was just a problem with the size of the kernel width. A width that
is too big is going to cause the coefficient matrix to become unstable.
This means that even the farthest points are going to contribute to the
total result, and the estimate for the cutoff will be wrong.